+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-2200**
   - Add Granule List page. Use Granule Inventory reports for generating granule CSVs
 
+- **CUMULUS-2225**
+  - Rename "Type" column "Error Type" on Failed Granules and Home pages and allow sorting
+  - Add Error Type filter to Failed Granules page
+
 - **CUMULUS-2218**
   - Add `Bulk Reingest` to Bulk Granules actions